connection timeout
my honeypot had over 1000 threads due to bots that do not close the connection and leave it open. ssh-honeypot creates a child process for every connection and does not set a timeout. the solution is kinda hacky but it works, i tried SSH_OPTIONS_TIMEOUT but it had no effect and due to the blocking nature of libssh, it would have taken a recode to callbacks.